pyproject.toml
[tool.poetry]
name = "auto-scheduler"
version = "0.1.0"
description = "Vietnam Hearts Scheduling Assistant"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
alembic = "1.12.1"
google-auth = "2.23.4"
google-auth-oauthlib = "1.1.0"
google-auth-httplib2 = "0.2.0"
google-api-python-client = "2.108.0"
google-generativeai = "^0.3.0"
python-dotenv = "^1.0.0"
pydantic = "2.5.0"
python-multipart = "0.0.6"
apscheduler = "^3.10.4"
jinja2 = "^3.1.3"
psycopg2-binary = "^2.9.9"
supabase = "^2.3.3"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
black = "23.11.0"
flake8 = "6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 